Struct gns::GnsSocket

pub struct GnsSocket<'x, 'y, S: GnsDroppable> { /* private fields */ }

Wrapper around mutiple low-level pointers. The socket is generic over the lifetime of both GnsGlobal and GnsUtils as they are required during its lifetime. The generic state S is evolving from IsCreated to either IsClient or IsServer depending on the functions used. The drop implementation make sure that everything related to this structure is correctly freed. The user has a strong guarantee that all the embedded points has been validated and thus, the available operations over the socket are safe.

Implementations

impl<'x, 'y, S> GnsSocket<'x, 'y, S> where
S: GnsDroppable,

pub unsafe fn into_inner(self) -> *mut ISteamNetworkingSockets

pub fn utils(&self) -> &GnsUtils

impl<'x, 'y, S> GnsSocket<'x, 'y, S> where
S: GnsDroppable + IsReady,

pub fn get_connection_real_time_status(
&self,
GnsConnection: GnsConnection,
nb_of_lanes: u32
) -> GnsResult<(GnsConnectionRealTimeStatus, Vec<GnsConnectionRealTimeLaneStatus>)>

pub fn get_connection_info(
&self,
GnsConnection: GnsConnection
) -> Option<GnsConnectionInfo>

pub fn flush_messages_on_connection(
&self,
GnsConnection: GnsConnection
) -> GnsResult<()>

pub fn close_connection(
&self,
GnsConnection: GnsConnection,
reason: u32,
debug: &str,
linger: bool
) -> bool

pub fn poll_messages<const K: usize, F>(&self, message_callback: F) -> usize where
F: FnMut(&GnsNetworkMessage<ToReceive>),

pub fn poll_event<const K: usize, F>(&self, event_callback: F) -> usize where
F: FnMut(GnsConnectionEvent),

pub fn poll_callbacks(&self)

pub fn configure_connection_lanes(
&self,
GnsConnection: GnsConnection,
lanes: &[GnsLane]
) -> GnsResult<()>

pub fn send_messages(
&self,
messages: Vec<GnsNetworkMessage<ToSend>>
) -> Vec<Either<GnsMessageNumber, EResult>>

impl<'x, 'y> GnsSocket<'x, 'y, IsCreated>

pub fn new(global: &'x GnsGlobal, utils: &'y GnsUtils) -> Option<Self>

Initialize a new socket in IsCreated state.

pub fn listen(
self,
address: Ipv6Addr,
port: u16
) -> Result<GnsSocket<'x, 'y, IsServer>, ()>

Listen for incoming connections, the socket transition from IsCreated to IsServer, allowing a new set of server operations.

pub fn connect(
self,
address: Ipv6Addr,
port: u16
) -> Result<GnsSocket<'x, 'y, IsClient>, ()>

Connect to a remote host, the socket transition from IsCreated to IsClient, allowing a new set of client operations.

impl<'x, 'y> GnsSocket<'x, 'y, IsServer>

pub fn accept(&self, connection: GnsConnection) -> GnsResult<()>

Accept an incoming connection. This operation is available only if the socket is in the IsServer state.

impl<'x, 'y> GnsSocket<'x, 'y, IsClient>

pub fn connection(&self) -> GnsConnection

Return the socket connection. This operation is available only if the socket is in the IsClient state.