gns

Type Alias GnsResult

pub type GnsResult<T> = Result<T, EResult>;

Outcome of many functions from this library, basic type alias with steam [sys::EResult] as error. If the result is [sys::EResult::k_EResultOK], the value can safely be wrapped, otherwise we return the error.

Aliased Type

enum GnsResult<T> {
    Ok(T),
    Err(EResult),
}

Variants

Ok(T)

Contains the success value

Err(EResult)

Contains the error value