gns

Struct GnsSocket

pub struct GnsSocket<S> { /* private fields */ }

GnsSocket is the most important structure of this library. This structure is used to create client (GnsSocket<IsClient>) and server (GnsSocket<IsServer>) sockets via the GnsSocket::connect and GnsSocket::listen functions. The drop implementation make sure that everything related to this structure is correctly freed, except the GnsGlobal instance and the user has a strong guarantee that all the available operations over the socket are safe.

Implementations

impl<S> GnsSocket<S>
where S: IsReady,

pub fn get_connection_real_time_status( &self, GnsConnection: GnsConnection, nb_of_lanes: u32, ) -> GnsResult<(GnsConnectionRealTimeStatus, Vec<GnsConnectionRealTimeLaneStatus>)>

Get a connection lane status. This call is possible only if lanes has been previously configured using configure_connection_lanes

pub fn get_connection_info( &self, GnsConnection: GnsConnection, ) -> Option<GnsConnectionInfo>

pub fn flush_messages_on_connection( &self, GnsConnection: GnsConnection, ) -> GnsResult<()>

pub fn close_connection( &self, GnsConnection: GnsConnection, reason: u32, debug: &str, linger: bool, ) -> bool

pub fn poll_messages<const K: usize>( &self, message_callback: impl FnMut(&GnsNetworkMessage<ToReceive>), ) -> Option<usize>

pub fn poll_event<const K: usize>( &self, event_callback: impl FnMut(GnsConnectionEvent), ) -> usize

pub fn configure_connection_lanes( &self, GnsConnection: GnsConnection, lanes: &[GnsLane], ) -> GnsResult<()>

pub fn send_messages( &self, messages: Vec<GnsNetworkMessage<ToSend>>, ) -> Vec<Either<GnsMessageNumber, EResult>>

impl GnsSocket<IsCreated>

pub fn new(global: Arc<GnsGlobal>) -> Self

Initialize a new socket in IsCreated state.

pub fn listen( self, address: IpAddr, port: u16, ) -> Result<GnsSocket<IsServer>, ()>

Listen for incoming connections, the socket transition from IsCreated to IsServer, allowing a new set of server operations.

pub fn connect( self, address: IpAddr, port: u16, ) -> Result<GnsSocket<IsClient>, ()>

Connect to a remote host, the socket transition from IsCreated to IsClient, allowing a new set of client operations.

impl GnsSocket<IsServer>

pub fn accept(&self, connection: GnsConnection) -> GnsResult<()>

Accept an incoming connection. This operation is available only if the socket is in the IsServer state.

impl GnsSocket<IsClient>

pub fn connection(&self) -> GnsConnection

Return the socket connection. This operation is available only if the socket is in the IsClient state.